chore: fix most warnings

+++ src/LindaDB/public/ldb/lv/linda_tuple.hxx
@@ -305,12 +305,12 @@ namespace ldb::lv {
                 if (a._position > b._position) {
                     const auto diff = a._position - b._position;
                     // ...or fail screaming
-                    assert_that(diff < std::numeric_limits<difference_type>::max());
+                    assert_that(diff < static_cast<std::size_t>(std::numeric_limits<difference_type>::max()));
                     return static_cast<difference_type>(diff);
                 }
                 const auto diff = b._position - a._position;
                 // ...or fail screaming
-                assert(diff < std::numeric_limits<difference_type>::max());
+                assert(diff < static_cast<std::size_t>(std::numeric_limits<difference_type>::max()));
 
                 return -static_cast<difference_type>(diff);
             }
@@ -331,8 +331,8 @@ namespace ldb::lv {
             [[nodiscard]] constexpr bool
             valid_step(difference_type by) const noexcept {
                 if (!_owner) return false;
-                if (by < 0) return _position >= -by;
-                if (by > 0) return _owner->_size - _position >= by;
+                if (by < 0) return _position >= static_cast<std::size_t>(-by);
+                if (by > 0) return _owner->_size - _position >= static_cast<std::size_t>(by);
                 return true;
             }
